[clang] create C++ temporaries on CXXBindTemporaryExpr when appropriate
Summary:
Like `MaterializeTemporaryExpr`, `CXXBindTemporaryExpr` is used by clang
to signal that a temporary variable is needed to hold the result of a
(sub-)expression. But, unlike the former, the latter only requires a
C++ temporary to be created in *some* of the cases, namely when the
sub-expression is not performing initialisation for a known variable
already (in particular one already created by
`MaterializeTemporaryExpr`! Or an actual program variable being
assigned).

Previously, we were missing creating (and thus destroying) these
C++ temporaries when needed. This diff fixes that by registering
`CXXBindTemporaryExpr` as a source of C++ temporaries in CScope.ml, and
adding logic to destroy only the C++ temporaries that ended up being
actually used.

Reviewed By: skcho

Differential Revision: D30455859

fbshipit-source-id: baf7029eb