[pulse] watch out for accidental state list collapses
Summary:
In several models we had a problematic use of the `let<*>/let<+>`
operators coupled with returning several possible states: we need to be
careful not to collapse *all* these disjuncts to `[Error _]` when only
*one* of them is an error. This happens whenever we do

```
  let<*> astate1 = ... in
  let<*> astate2 = ... in
  [astate1; astate2]
```

This gives the wrong result `[Error _]` when either of these two failed,
instead of a two-state result with at least one error in it, eg
`[Error err1; astate2]`.

The correct way to write these whenever several states are to be
returned is to not use the `let<*>`/`let<+>` operators at the top level:
```
  let astate1 = let<*> x = ... in ... in
  let astate2 = ... in
  [astate1; astate2]
```

Concretely the incorrect behaviour made us miss some coverage and so
introduced false negatives.

Reviewed By: skcho

Differential Revision: D30541428

fbshipit-source-id: af04e4ccf