[erl-frontend] Add support for maps with approximate Pulse model
Summary:
This diff adds support for Erlang maps using an approximate Pulse model to store only the most recent key/value. The translation uses 4 built-in functions: `map_create`, `map_has_key`, `map_lookup`, `map_update` to translate the syntax primitives in the Erlang language. The translation itself is precise, but the underlying Pulse models for the 4 built-in functions only store the most recent key and value. The Pulse model uses an underapproximation: if we don't know if a key/value is present we stop that branch.

The diff also introduces two new errors to be reported: `badkey` (trying to access/update a non-existing key) and `badmap` (trying to access/update an expression that is not a map).

Reviewed By: rgrig

Differential Revision: D30192257

fbshipit-source-id: 20bf5cc80