[SelfInBlock] Change domain to set of memory states
Summary:
This diff changes domain to set of memory states in order to distinguish them by branches. For example,

```
if (b) {
  [weakSelf foo];
} else {
  [weakSelf bar];
}
```

Before the diff, it collected all `weakSelf` uses as one memory state, i.e. a set. Thus, it reported a FP issue that said `weakSelf` was used multiple times.

After this diff, it collects the states separately, so now it knows `weakSelf` is used once on each path.

About state explosion issue: Note that the memory states are abstracted by

* `weakSelf` usages and
* `strongSelf` assignments.

so the number of memory states should be reasonably small for ordinary ObjC blocks, in which `weakSelf` is used small number of time and `strongSelf` is assigned small number of time.

Reviewed By: ngorogiannis

Differential Revision: D30869730

fbshipit-source-id: baf7d08d8