[ConfigImpact] Ignore unknown setter/getter in strict mode
Summary: In many cases, simple setter or getter does not call other methods, so they usually have empty summary. A problem is when the body of setter/getter is unknown, e.g. they are declared in an interface. At that time, the strict mode addressed them as leaf functions, thus introduced some FPs.  This diff heuristically checks the method name and if it has prefix of "set" or "get", it ignores the function call.

Reviewed By: ezgicicek

Differential Revision: D30962209

fbshipit-source-id: d91341c0b