[config] compute [is_originator] correctly
Summary:
`CommandLineOption.is_originator` is supposed to be true whenever the current process is the root of the infer process tree. This is done by checking for the presence of an environment variable.

In remote execution environments such as under `buck` this variable does not translate, so Infer processes invoked by buck will have `is_originator=true` which leads to bugs (logging wrong info, not cleaning caches).

This diff computes the right value and moves it to `Config`.

Reviewed By: martintrojer

Differential Revision: D31020205

fbshipit-source-id: 08f87c2de