[erl-lineage] Checker that produces lineage graphs
Summary:
For each procedure, compute a summary that is a graph whose vertices are
pairs (N,V) where N is a position in the program and V is a variable. A
position in the program is an instruction in a node in the flowgraph.

Here is an example:

```
  4 f() ->
  5   X = 1,
  6   Y = g(X,X), % Local(X,5)->ArgumentOf(0,g); Local(X,5)->ArgumentOf(1,g); ReturnOf(g)->Local(Y,6); Local(X,5)=>Local(Y,6)
  7   Y.          % Local(Y,6)->Return
  8
  9 g(A, B) ->    % Argument(0)->Local(A,9); Argument(1)->Local(B,9)
 10   AA = h(A),  % Local(A,9)->ArgumentOf(0,h); ReturnOf(h)->Local(AA,10); Local(A,9)=>Local(AA,10)
 11   BB = h(B),  % Local(B,9)->ArgumentOf(0,h); ReturnOf(h)->Local(BB,11); Local(B,9)=>Local(BB,11)
 12   AA + BB.    % Local(AA,10)->Return; Local(BB,11)->Return
 13
 14 h(Z) ->   % Argument(0)->Local(Z,14)
 15   3 * Z.  % Local(Z,14)->Return
```

In the example, I used line numbers for location: in implementation those are (flowgraph node, instruction index). A "data" like `Local(X,5)` means "the value of X after the statement at location 5"; "Argument 0" means "the value of the first argument at the beginning of the procedure"; "Return" means "the returned value"; "ArgumentOf(0,h)" means "the value of the first argument at the beginning of procedure h". Note that edges with "ArgumentOf" targets introduce imprecision, because two calls to the same function get merged -- this is the lack of context-sensitivity. The "summary" edges mentioned below alleviate this problem a bit.

There are two types of flow: a direct flow `x->y` means roughly that `x` is the result of some simple computation involving `y` (e.g., evaluating an expression that contains `y` and assigning that value to `x`); a summary flow `x=>y` means that `x` is returned by a callee that used the value `y` (at least, the summary of the callee says that the value of `y` is used).

Reviewed By: ezgicicek

Differential Revision: D30896651

fbshipit-source-id: f88f9e979