[sledge] Rework first-order solver
Summary:
Before this diff the interaction between interpreted and uninterpreted
terms in the equality solver was incorrect in some cases. One source
of issues is that applying a solution substitution to an interpreted
term can produce an uninterpreted (aka solvable) one, and vice
versa. This possibility was not accounted for in all cases, leading to
invariant violations followed by non-termination and incompleteness.

The main driver for these changes is needing a more nuanced
interpretation of the "use lists", which are used when adding a new
equation to find additional equations that might be consequences.  The
use map itself becomes sparser, (O(|equality partitions|) instead of
O(|total terms|)), and the notion of "use" takes maximal solvable
subterms into account, as well as operating modulo the representatives
map.

Before this change:

Map each solvable in the carrier to its immediate super-terms.

After this change:

Map solvable representatives to their "uses". A "use" of a term is
similar to an immediate super-term, but modulo the representatives
map:

- Every use of solvable `a'` has a solvable subterm whose
  representative's solvables contain `a'`: if `f(b₀,…,bₙ) ∈ use(a')`,
  then `∃i. a' ∈ solvables(rep(bᵢ))`.  Also `solvables(f(b₀,…,bₙ)) ⊆
  dom(rep) ∪ dom(cls)`.

- Every representative is congruent to some use of the representative
  of each of its solvable subterms with solvable reps: if `f(b₀,…,bₙ)
  ∈ rng(rep)` then for each `i`, if `rep(bᵢ)` is solvable, then `∃
  f(c₀,…,cₙ) ∈ use(rep(bᵢ))` such that `rep(cⱼ) = rep(bⱼ)` for
  `0≤j≤n`.

The representatives map is also changed to be sparser, now not
including identity mappings for interpteted representatives or atomic
terms, which enables a significant simplification of quantifier
elimination support functions.

Reviewed By: peterogithub

Differential Revision: D31129952

fbshipit-source-id: 81081284e