[erl-frontend] Support list append up to configurable depth
Summary: Add support for list append (both `++` operator and `lists:append/2`) with an approximate Pulse model that supports abstract lists up to a configurable depth. Furthermore, the `--erlang-reverse-unfold-depth` argument is renamed to `--erlang-list-unfold-depth` as it now applies to both `reverse` and `append` (and can be potentially applied to other list functions in the future).

Reviewed By: rgrig

Differential Revision: D31271930

fbshipit-source-id: 5e3d14234