[sledge] Normalize Sh wrt context as part of Sh.simplify
Summary:
This diff adds a full normalization of the Sh formula with respect to
its first-order context to the simplification operation. This can
reduce the size of the representation further than simplification
alone since e.g. distinct disjuncts can normalize to the same
formula. This normalization has a nontrivial computational cost, and
in some cases the reduction leads to less work later and is worth the
cost, while in other cases there is not enough reduction to pay for
the normalization. So for the time being this normalization is enabled
only with a command line flag.

Differential Revision: D31283436

fbshipit-source-id: 353d9c993