[kotlin] do not die on Hil reassignment
Summary:
`Bindings` keeps a mapping from variables to expressions for use in `Hil` analysers. Variables are supposed to be in single-static assignment form but this sometimes fails to be the case.

The Kotlin code in the provided test leads to a crash when `--debug-exceptions` is set. This diff logs an error but continues execution, in line with what happens when `--debug-exceptions` is not set.

Reviewed By: jvillard

Differential Revision: D31377654

fbshipit-source-id: adac4b21c