[erl-frontend] Fix return value handling in expression translation
Summary: When translating an expression, if the result in the environment is not an identifier (but for example a var), we generate a fresh identifier. However, we did not update the environment. Most of the expressions still worked because for convenience they received the new identifier directly, but some of them (such as `case`, `block`, `if` did not).

Reviewed By: rgrig

Differential Revision: D31392706

fbshipit-source-id: 09f290c8d