[erl-frontend] Add support for dispatching BIFs in Pulse plus model for is_map
Summary: Set up infrastructure to be able to dispatch Erlang BIFs (built-in functions) to Pulse models and add support for `erlang:is_map/1` BIF. Previously we were assuming that if a function is not imported, then it is in the local module. This is not true, because the functions in the `erlang` module can be used unqualified. So now we store not only the imported functions, but also the locals, and if a function does not match those, we assume it's built-in.

Reviewed By: rgrig

Differential Revision: D31500811

fbshipit-source-id: 362d86773