[pulse] improve treatment of unknown functions
Summary:
When Pulse encounters a call to a function for which we have no summary,
it doesn't know what the effect of this function is and so risks
reporting subsequent false positives which in turn may mask further
true positives unrelated to the unknown call. To mitigate this,
following in the biabduction analysis footsteps, we try to give these
unknown calls a semantics that is as "angelic" as possible and scramble
the values returned by, and passed by pointer to, the call.

This is difficult to do in a way that works interprocedurally, i.e. when
the unknown function call happens in a callee of the current procedure:
the state when analysis the callee may be a sub-set of the state at the
call site and then we don't scramble enough of the state as a result. To
mitigate *that*, we scramble all the fields of the target type of a
pointer, according to the tenv. However, this assumes the type of the
pointer is correct and no other field will ever be accessed. This is not
true, especially in the case of models, where the fields used by a Pulse
model may be completely different than the ones used in the actual
implementation, causing issues.

Instead, we can do without the type environment and remember which
locations have been passed to unknown functions. When applying a callee
summary, we replay the havoc'ing of fields for those fields that are in
the caller but not in the callee. This is more flexible than the
previous approach and ensures that the effect of unknown calls is the
same intra-procedurally and inter-procedurally. Also, there's no reason
to stop at the first level when havoc'ing so let's havoc every access
reachable from the parameters (just like we free/initialise all the
addresses reachable from the parameters).

Reviewed By: skcho

Differential Revision: D31056813

fbshipit-source-id: 203ef8f15