[sledge][test] Sort tests by decreasing runtime in perf mode
Summary:
Measurement noise in the longer-running tests decreases, and overall
test time decreases, if test are sorted in decreasing order by
baseline runtime. The hypothesis for this is that, without the ability
to disable CPU frequency scaling and pin processes to cores, running
the heavy tests first leads to them being run in a context that is
more constantly heavily-loaded than randomizing the order. The
shorter-running tests see more noise, but it matters less in terms of
decision making based on the performance results.

Differential Revision: D31457333

fbshipit-source-id: c082d7641