[cost] models for C++ iterators
Summary:
This diff adds models for function iterators such as `end`, `begin`, `cbegin` and `cend`. Loop comprehensions such as

```
 for (auto i : vec) {
  }
```
are also covered because they are translated to `begin` and `end`.

While creating the models for begin and end, we found out that the pruning between the values involved in the `operator!=` was not correct. Pruning works well when the prune statement includes the full expression `PRUNE(n$0 < n$1, true)`. However, for `n$11=_fun_std::operator!=(&itt, &temp)` where `temp` is the vale of `vec.end`, the prune statement would be `PRUNE(n$11, true)`. Therefore, `n$11` would be pruned to 1 instead to the comparison between `itt` and `temp`.

This diff fixes that issue by:
1) Extending the BufferOverrunDomain (aka Inferbo's memory domain) to include a new alias called `CppIteratorCmp` to store the relationship between `n$11`, `itt` and `temp`.
2) Updating the `CppIteratorCmp` alias in the `iterator_ne` function.
3) Extending the `prune_helper` function to update the prune mem with the information of the new alias. Match is done based on the register name (`n$11`).

Diff co-authored with skcho. Thanks for the help!

Reviewed By: ezgicicek, skcho

Differential Revision: D31433265

fbshipit-source-id: 628372a1e