[pulse] Create a non-disjunctive domain along with the disjunctive one
Summary:
This diff creates a non-disjunctive domain that will keep track of a more global state along with the disjunctive domain.

Now, Pulse's domain is transformed from a list of disjuncts to a pair of a list of disjuncts and non-disjunctive domain.

Note that, unlike having a completely separate domain with its own transfer function,  the advantage of the non-disjunctive domain is that it can make use of the disjuncts when executing instructions which would not be possible otherwise.

This change requires changing Pulse's underlying transfer functions in `AbstractInterpreter`: e.g.
 1) join now must also join the non-disj domain
 2) while executing the instruction non-disjunctive domain is threaded through

Initially, non-disjunctive domain is void and useless. It will be changed/used for unnecessary copy checker in the upcoming diffs.

Reviewed By: skcho

Differential Revision: D31731427

fbshipit-source-id: 318e380c4