[pulse] Unnecessary checker with Copy non-disj domain
Summary:
This diff adds a map of vars to whether they are copied/modified as a non-disj domain for the unnecessary checker.
- whenever we see a call to a copy constructor, we add the variable as `Copied` with its heap snapshot
- at the time of exit scope, we check whether  a copy is modified, if so we mark it as `Modified`
- when joining states, `Copied < Modified`, so when something is modified only on one path we  won't report an unnecessary copy.

There are three possible states for a var:
- no copy created, hence var has no binding in the map
- Copied :  var is coming from a copy via copy constructor
- Modified: var which was initially copied is now modified

## Limitations

### Changes to source after copy
Note that at this time, we do not check whether the source is modifed. If source is modified after the copy, it might not make sense to point out the unnecessary copy. This is a further check which can be implemented later.

```
void source_copy_mod_ok(bool b){
  Foo source;
  auto copy = source;
  source.x = 2;
  read_x(copy);
}

```
### Reporting everywhere

We currently report in all functions, but ideally we might not want to report on constructors or class initializers. This can be fixed later.

### Other patterns

There are a bunch of more patterns that copy stuff: e.e. copy assignment, `optional::value_or`, `follly:get_default` etc. We do not currently handle them but should be easy to extend the analysis.

Reviewed By: skcho

Differential Revision: D31757424

fbshipit-source-id: a7b47d399