[paramNotNull] Taking store into account in the Parameter Not Null Checked checker
Summary: This adds support for assignments in the Parameter Not Null Checked check. The idea is that when the block is assigned to something, our information about whether it's checked for null gets removed, and if it is assigned to another block that had been checked for null before (or is annotated), we then take that information into account.

Reviewed By: skcho

Differential Revision: D32136724

fbshipit-source-id: 4503dffea