[tests] Putting retain cycle test in own folder
Summary:
Organising biabduction tests such that we can migrate them to Pulse and/or delete them when the corresponding biabduction checker has been migrated.
The test I moved here was a retain cycle test but was written in a way that we would get an NPE if we didn't get a retain cycle. So I made it more in sync with the current formats and moved it to own retain cycle sub-folder.

Reviewed By: jvillard

Differential Revision: D32204850

fbshipit-source-id: b1203b705