[tests] [5/n] Move test about initialization of structs from biabduction to Pulse
Summary:
Biabduction will be phased out soon, so tests about the frontend that rely on biabduction should be migrated to Pulse.
In this case the test was relying on divide_by_zero from biabduction, now we use NULLPTR_DEREFERENCE from Pulse. Moreover, it was in the objc section, but the code is just c code, so moved it to the c section of Pulse.

Reviewed By: ngorogiannis

Differential Revision: D32248117

fbshipit-source-id: c0a77eb0b