[simple-lineage] Track constants
Summary:
In the lineage graph, variables are replaced by "terms", which can be
variables or constants. For now, we track integer constants and string
constants. (Atoms appear as string constants because of the way they are
currently modelled in the Erlang frontend. We should have some special
treatement for them.)

Differential Revision: D32563349

fbshipit-source-id: c0ce50f934