[inferbo][cost] Improve iterator comparison alias mechanism
Summary:
In D31433265 (https://github.com/facebook/infer/commit/9dd7954bf7b58e2c11e6f7886c51c9e816bbe4ed), we had heuristically assumed that when we have ( ` iter_lhs != iter_rhs `), `iter_rhs` would have the value of the end of the iterator. This is indeed the case in container comprehensions but not always. For instance, consider the added example program where we check if `it != container.begin()`. Currently we raise an exception in the analysis due to getting the bound of a negative value because we are pruning the iterator as [-oo, -1] by the wrong assumption that rhs of the comparison is an iterator end (hence pruned by `<`). This is wrong. We should prune it with `>` because it is originating from begin.

This diff fixes the issue by considering both cases where RHS could be either `begin` or `end`. To do that ,we keep an additional mapping:
- whenever we call `iter.begin()` or `iter.end()`, we keep track of whether `iter` was a begin or end
- whenever we call ` iter_rhs != iter_rhs`, we lookup the begin/end mapping for `iter_rhs`. If it is `begin`, we prune by `>` and if it is `end` we prune by `<` (as before).

Reviewed By: skcho

Differential Revision: D32617577

fbshipit-source-id: 4292a07b4b